mlops/requirements.txt
 
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
mlflow>=2.8.0
# MLflow avec support PostgreSQL
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
alembic>=1.12.0
feast>=0.30.0 
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
 